feat(engine): sparse 확장 — ImplicitOccupancy(복셀화 폐기·64비트 키·온디맨드 클리어런스)로 정밀 셀 대응
셀을 줄이면 셀 수가 부피로 폭증해 점유배열(25mm 130MB·10mm 2GB)·클리어런스 거리맵(520MB)·
A* int 키(10mm 오버플로)가 동시에 터지던 문제를 근본 해결(S1~S4).

- S3 box_index.hpp: 장애물 AABB를 유니폼그리드(broadphase)로 색인 → is_blocked/최근접질의.
  메모리 O(장애물 수), 셀 크기와 무관. 반열린 중첩으로 Dense 복셀화와 동일 점유(O1).
- S1/S3 ImplicitOccupancy: 박스색인 + 깔린배관 해시셋. Dense vector<uint8_t> 폐기.
- S2 astar_weighted: state_of/unlin 64비트 → 20.3억 셀(10mm)에서 오버플로 없음.
- S4 CostModel: HasClearanceQuery 컨셉 + if constexpr 온디맨드 클리어런스(전역 거리변환 폐기).
- capi route_multi: 격자>5M셀이면 Implicit 자동 전환, 이하 Dense(골든 바이트 불변).
- C# BuildEngineForRows: 대형격자 클리어런스 재활성(온디맨드라 저렴).

검증: ctest 10/10(신규 implicit = Dense==Implicit 전수 일치, golden/capi 불변),
25mm ALKA 14/14·clearance ON·크래시 0, 10mm ALKA exit 0(이전 exit 255 크래시 해소).

+++ cpp/include/routing3d/astar.hpp
@@ -170,9 +170,9 @@ AStarResult astar_weighted(const Occ& occ, Cell start, Cell goal, const RoutePar
     }
 
     CostModel<Occ> model(occ, params, corridor);
-    auto state_of = [&](int lin, int dir) -> long long {
-        return static_cast<long long>(lin) * 7 + (dir + 1);
-    };
+    // 상태키 = lin*7 + (dir+1). lin 은 백엔드 키 타입(Dense/Sparse=int, Implicit=long long)을
+    // long long 으로 받아 곱한다 → 10mm 거대격자(20억 셀)에서도 int 오버플로 없음(S2).
+    auto state_of = [&](long long lin, int dir) -> long long { return lin * 7 + (dir + 1); };
 
     std::priority_queue<detail::PQItem, std::vector<detail::PQItem>, detail::PQCmp> open;
     std::unordered_map<long long, double> g;
@@ -204,11 +204,11 @@ AStarResult astar_weighted(const Occ& occ, Cell start, Cell goal, const RoutePar
         if (cur.cell == goal) {
             std::vector<Cell> path;
             long long s = st;
-            path.push_back(occ.unlin(static_cast<int>(s / 7)));
+            path.push_back(occ.unlin(s / 7));  // unlin 은 long long 인자(백엔드 무관, S2).
             auto it = came.find(s);
             while (it != came.end()) {
                 s = it->second;
-                path.push_back(occ.unlin(static_cast<int>(s / 7)));
+                path.push_back(occ.unlin(s / 7));
                 it = came.find(s);
             }
             for (size_t a = 0, b = path.size() - 1; a < b; ++a, --b) std::swap(path[a], path[b]);
